img-processor/requirements.txt
fastapi
uvicorn
loguru
argparse
opencv-python
ffmpeg-python
graphviz
numpy
boto3
python-dotenv
mux_python
logger
openai